requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.7
emailmessage==0.2.1
et_xmlfile==2.0.0
invoke==3.0.3
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
paramiko==4.0.0
pycparser==3.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
